*: move download function to shared
This moves the DownloadFileSha256 function and its dependent code to
shared.
