+++ client/util.go
@@ -1,16 +1,12 @@
 package lxd
 
 import (
-	"crypto/sha256"
-	"fmt"
 	"io"
 	"net"
 	"net/http"
 	"net/url"
 
 	"github.com/lxc/lxd/shared"
-	"github.com/lxc/lxd/shared/cancel"
-	"github.com/lxc/lxd/shared/ioprogress"
 )
 
 func tlsHTTPClient(client *http.Client, tlsClientCert string, tlsClientKey string, tlsCA string, tlsServerCert string, insecureSkipVerify bool, proxy func(req *http.Request) (*url.URL, error)) (*http.Client, error) {
@@ -82,64 +78,6 @@ func unixHTTPClient(client *http.Client, path string) (*http.Client, error) {
 	}
 
 	return client, nil
-}
-
-func downloadFileSha256(httpClient *http.Client, useragent string, progress func(progress ProgressData), canceler *cancel.Canceler, filename string, url string, hash string, target io.WriteSeeker) (int64, error) {
-	// Always seek to the beginning
-	target.Seek(0, 0)
-
-	// Prepare the download request
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return -1, err
-	}
-
-	if useragent != "" {
-		req.Header.Set("User-Agent", useragent)
-	}
-
-	// Perform the request
-	r, doneCh, err := cancel.CancelableDownload(canceler, httpClient, req)
-	if err != nil {
-		return -1, err
-	}
-	defer r.Body.Close()
-	defer close(doneCh)
-
-	if r.StatusCode != http.StatusOK {
-		return -1, fmt.Errorf("Unable to fetch %s: %s", url, r.Status)
-	}
-
-	// Handle the data
-	body := r.Body
-	if progress != nil {
-		body = &ioprogress.ProgressReader{
-			ReadCloser: r.Body,
-			Tracker: &ioprogress.ProgressTracker{
-				Length: r.ContentLength,
-				Handler: func(percent int64, speed int64) {
-					if filename != "" {
-						progress(ProgressData{Text: fmt.Sprintf("%s: %d%% (%s/s)", filename, percent, shared.GetByteSizeString(speed, 2))})
-					} else {
-						progress(ProgressData{Text: fmt.Sprintf("%d%% (%s/s)", percent, shared.GetByteSizeString(speed, 2))})
-					}
-				},
-			},
-		}
-	}
-
-	sha256 := sha256.New()
-	size, err := io.Copy(io.MultiWriter(target, sha256), body)
-	if err != nil {
-		return -1, err
-	}
-
-	result := fmt.Sprintf("%x", sha256.Sum(nil))
-	if result != hash {
-		return -1, fmt.Errorf("Hash mismatch for %s: %s != %s", url, result, hash)
-	}
-
-	return size, nil
 }
 
 type nullReadWriteCloser int
